backend/pyproject.toml
[project]
name = "second-brain-backend"
version = "0.1.0"
description = "AI-Powered Personal Knowledge System (Second Brain)"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn==0.30.0",
    "sqlalchemy==2.0.30",
    "psycopg[binary]==3.2.9",
    "celery==5.4.0",
    "redis==5.1.1",
    "PyJWT==2.9.0",
    "bleach==6.1.0",
    "psutil==6.1.0",
    "chromadb==0.5.0",
    "networkx==3.3",
    "pydantic-settings==2.3.0",
    "httpx==0.27.0",
    "beautifulsoup4==4.12.0",
    "youtube-transcript-api==0.6.0",
    "yt-dlp==2024.12.13",
    "python-multipart==0.0.9",
    "numpy<2.0.0",
]

[project.optional-dependencies]
desktop = [
    "pyinstaller==6.11.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
